styling hero details page
