+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
 import './App.css';
-import HeroDetails from "./components/HeroDetails";
+import HeroDetails from "./components/HeroDetails/HeroDetails";
 import Home from "./layout/Home/home";
 
 function App() {
