+++ frontend/src/components/HeroDetails/style.css
@@ -1,0 +1,20 @@
+.container--hero-details {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.hero-details{
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-self: flex-start;
+  margin: 0px 10px
+}
+
+.hero-details p {
+  margin-block-start: 0em;
+  margin-block-end: 0em;
+}
